containers/slo-tools/requirements.txt
google-cloud-bigquery==1.25.0
google-cloud-storage==1.29.0
